+++ src/sqlfluff/dialects/dialect_postgres.py
@@ -58,6 +58,9 @@ postgres_dialect.add(
     JsonOperatorSegment=NamedSegment.make(
         "json_operator", name="json_operator", type="binary_operator"
     ),
+    DollarQuotedLiteralSegment=NamedSegment.make(
+        "dollar_quote", name="dollar_quoted_literal", type="literal"
+    ),
 )
 
 
@@ -68,6 +71,8 @@ postgres_dialect.replace(
             Sequence(OneOf("IGNORE", "RESPECT"), "NULLS", optional=True),
             Ref("OverClauseSegment"),
         ),
+        # Filter clause supported by both Postgres and SQLite
+        Ref("FilterClauseGrammar"),
     ),
     BinaryOperatorGrammar=OneOf(
         Ref("ArithmeticBinaryOperatorGrammar"),
@@ -78,6 +83,22 @@ postgres_dialect.replace(
         Ref("JsonOperatorSegment"),
     ),
 )
+
+
+@postgres_dialect.segment(replace=True)
+class FunctionDefinitionGrammar(BaseSegment):
+    """This is the body of a `CREATE FUNCTION AS` statement."""
+
+    match_grammar = Sequence(
+        "AS",
+        OneOf(Ref("QuotedLiteralSegment"), Ref("DollarQuotedLiteralSegment")),
+        Sequence(
+            "LANGUAGE",
+            # Not really a parameter, but best fit for now.
+            Ref("ParameterNameSegment"),
+            optional=True,
+        ),
+    )
 
 
 @postgres_dialect.segment(replace=True)